refactor(bigquery): move synthetic write message to model_ext
